Classifying Thyroid Disease Using a Multi-Layer Perceptron/requirements.txt
streamlit
numpy
matplotlib
google-generativeai
python-dotenv
